+++ src/checks.rs
@@ -5,6 +5,7 @@ mod incorrect_delimiter;
 mod key_without_value;
 mod leading_character;
 mod lowercase_key;
+mod quote_character;
 mod space_character;
 mod unordered_key;
 
@@ -21,6 +22,7 @@ fn checklist() -> Vec<Box<dyn Check>> {
         Box::new(leading_character::LeadingCharacterChecker::default()),
         Box::new(key_without_value::KeyWithoutValueChecker::default()),
         Box::new(lowercase_key::LowercaseKeyChecker::default()),
+        Box::new(quote_character::QuoteCharacterChecker::default()),
         Box::new(space_character::SpaceCharacterChecker::default()),
         Box::new(unordered_key::UnorderedKeyChecker::default()),
     ]